Organize backend into com.chatbot package
