+++ chatbot/Backend/src/com/chatbot/ChatUI.java
@@ -1,3 +1,5 @@
+package com.chatbot;
+
 import java.io.*;
 import jakarta.servlet.*;
 import jakarta.servlet.annotation.*;
