+++ chatbot/run.bat
@@ -12,10 +12,10 @@ if not exist API_KEY.txt (
 
 echo Compiling Java files...
 set CLASSPATH=./Backend/lib/*
-set JAVAFILES=./Backend/src/*.java
+set JAVAFILES=./Backend/src/com/chatbot/*.java
 
 javac -cp "%CLASSPATH%" %JAVAFILES% -d ./Backend/bin
 
 REM Run server
 echo Starting server...
-java -cp "./Backend/bin;%CLASSPATH%" ServletMain
+java -cp "./Backend/bin;%CLASSPATH%" com.chatbot.ServletMain
